+++ Siesa.SDK.Shared/Entities/E00041_ModuleFeature.cs
@@ -24,7 +24,7 @@ namespace Siesa.SDK.Entities
 
 		[ForeignKey("Module")]
 		[Required]
-		public int RowidModule { get; set; }
+		public short RowidModule { get; set; }
 
 		[ForeignKey("Feature")]
 		[Required]
